feat: separate functions for tree-shaking
